% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{alert}
\alias{alert}
\title{\code{alert}}
\usage{
alert(message)
}
\arguments{
\item{message}{a message to send}
}
\value{
Send an alert dialog to the browser
}
\description{
Send an alert dialog to the browser
}
\examples{
alert("hello world")
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Window/alert}
}
\keyword{alert}
\keyword{browsertools,}
