% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_js_object.R
\name{as_js_object}
\alias{as_js_object}
\title{\code{as_js_object}}
\usage{
as_js_object(x)
}
\arguments{
\item{x}{a data frame object}
}
\value{
Restructures a data.frame to JavaScript Object
}
\description{
Restructures a data.frame to JavaScript Object
}
\examples{
df <- as_js_object(x = iris)
console_log(df)
}
\keyword{browsertools,}
\keyword{object}
