% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{clear_input}
\alias{clear_input}
\title{\code{clear_input}}
\usage{
clear_input(elem, value = NULL)
}
\arguments{
\item{elem}{the id or class name of an html element}
}
\value{
Resets an html inputs value
}
\description{
Resets an html inputs value
}
\examples{
clear_input(elem = "#mydiv")
}
\keyword{browsertools,}
\keyword{clear}
\keyword{input,}
