% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{console_error}
\alias{console_error}
\title{\code{console_error}}
\usage{
console_error(message)
}
\arguments{
\item{message}{an error to display}
}
\value{
Sends an error message to the browser's console
}
\description{
Sends an error message to the brower's console
}
\examples{
console_error(message = "Error: 'object' undefined")
}
\keyword{browsertools,}
\keyword{console}
\keyword{debugging,}
