% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_warn.R
\name{console_warn}
\alias{console_warn}
\title{\code{console_warn}}
\usage{
console_warn(message)
}
\arguments{
\item{message}{a message to display}
}
\value{
Outputs a warning message to the console
}
\description{
Send a warning message to the console
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Display Warning Message in the Dev Console"),
    tags$p("Open the browser's dev console and click the button below."),
    tags$button(
       id = "sendWarning",
       class = "shiny-bound-input action-button",
       "Send Warning"
    )
  )
  server <- function(input, output, session) {
    observeEvent(input$sendWarning, {
        browsertools::console_warn(
            message = "This is a warning message"
        )
    })
  }
  shinyApp(ui, server)
}
}
\references{
\preformatted{ \url{https://developer.mozilla.org/en-US/docs/Web/API/Console/warn}
}
}
\seealso{
\code{\link[=console_error]{console_error()}}, \code{\link[=console_log]{console_log()}}, \code{\link[=console_table]{console_table()}}
}
\keyword{browsertools}
\keyword{console}
\keyword{warn}
