% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug}
\alias{debug}
\title{\code{Debug}}
\usage{
debug()
}
\value{
Print JavaScript errors in the R console
}
\description{
Print JavaScript errors in the R console. This function
is designed to be used within the Shiny Server
}
\examples{
if (interactive()) {
  ui <- tagList(
     tags$h2("Browsertools demo"),
     tags$p("JavaScript messages will print in the R Console")
  )
  server <- function(input, output, session) {
     browsertools::debug()
  }
  shinyApp(ui, server)
}
}
\keyword{browsertools,}
\keyword{debug,}
\keyword{error}
