% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{get_shiny_session}
\alias{get_shiny_session}
\title{\code{get_shiny_session}
A function for finding the shiny session}
\usage{
get_shiny_session()
}
\description{
\code{get_shiny_session}
A function for finding the shiny session
}
\keyword{browsertools,}
\keyword{session}
