% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{hidden}
\alias{hidden}
\title{\code{hidden}}
\usage{
hidden(..., css = "browsertools-hidden")
}
\arguments{
\item{...}{A Shiny tag element or a tagList object}

\item{css}{An optional class name used to hide elements
(Default class = "browsertools")}
}
\value{
Hide elements by when rendered
}
\description{
Hide an element or elements by default (i.e., at the time of page render)
}
\examples{
if (interactive()) {
 hidden(tags$p("hello, world"), tags$p("this is a test"))
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Hidden Elements"),
    tags$p("This element is visible by default"),
    browsertools::hidden(
      tags$p(
        id = "myHiddenElement",
        "This element is hidden by default"
      )
    )
  )
  server <- function(input, output, session) {
  }
  shinyApp(ui, server)
}

}
\keyword{browsertools}
\keyword{hidden}
