% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_html.R
\name{inner_html}
\alias{inner_html}
\title{\code{inner_html}}
\usage{
inner_html(elem, content, append = FALSE, delay = NULL)
}
\arguments{
\item{elem}{the ID or class name of an html element}

\item{content}{an html string, Shiny tag, or Shiny tag list}

\item{append}{an option that appends value of string or replaces it}

\item{delay}{an optional arg to add a brief pause
before sending the content to the html element.
Ideal for content that is rendered by the server.
Input is time in milliseconds.}
}
\value{
Update the content of an element by id or class name
}
\description{
Update the content of an element by id or class name
}
\examples{
if (interactive()) {
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Inner Html Example"),
    tags$p("Click the button to update the following message."),
    tags$p(
      id = "textToUpdate",
      "[This text will be updated]"
    ),
    tags$button(
      id = "updateText",
      class = "shiny-bound-input action-button",
      "Update Text"
    )
  )
  server <- function(input, output, session) {
    browsertools::inner_html(
      elem = "#textToUpdate",
      content = tags$strong("This is a bold message!")
    )
  }
  shinyApp(ui, server)
}
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Element/innerHTML}
}
\seealso{
\code{\link[=inner_text]{inner_text()}}, \code{\link[=insert_adjacent_html]{insert_adjacent_html()}}
}
\keyword{innerhtml}
