% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{inner_text}
\alias{inner_text}
\title{\code{inner_text}}
\usage{
inner_text(elem, string, delay = NULL)
}
\arguments{
\item{elem}{an element to select (e.g., ID, class, tag, etc.)}

\item{string}{a string used to insert into the element}

\item{delay}{an optional arg that adds a brief pause before
sending the content to the html element. Ideal
for content that is rendered server-side. Input
time is in milliseconds.}
}
\value{
Modify the text of an element
}
\description{
Modify the text of an element
}
\examples{
inner_text(elem = "#mydiv", string = "Hello, world", delay = 200)
}
\keyword{browsertools,}
\keyword{innertext}
