% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_elem.R
\name{print_elem}
\alias{print_elem}
\title{\code{print_elem}}
\usage{
print_elem(elem)
}
\arguments{
\item{elem}{an element to select (i.e., ID, class, tag, etc.)}
}
\value{
Find and print an HTML element in the R console.
}
\description{
Find and print an HTML element in the R console. This function is designed
to be used during application development rather than production
applications as the corresponding JavaScript does not create a new
instance when called. Use this function to debug the HTML markup of
elements or simply to view a structure of an element. This may be
useful when applications are running the viewer pane rather than in the
browser.
}
\examples{
if (interactive()) {
   ui <- tagList(
       browsertools::use_browsertools(),
       tags$div(
         id = "mydiv",
         tags$h2("My Element"),
         tags$p(
           "This is an example element. The structure will be printed",
           "in the R console."
         )
       )
   )
   server <- function(input, output) {
     observe({
        print_elem(elem = "#mydiv")
     })
   }
}

}
\seealso{
\code{\link[=debug]{debug()}}
}
\keyword{browsertools}
\keyword{debugging}
\keyword{print}
