% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{refresh_page}
\alias{refresh_page}
\title{\code{refresh_page}}
\usage{
refresh_page()
}
\value{
Trigger a page refres
}
\description{
Trigger a page refresh
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
     browsertools::refresh_page(),
     tags$h2("Refresh App Example"),
     tags$p("Click the button to refresh the app"),
     tags$button(
         id = "refreshApp",
         class = "shiny-bound-input action-button",
         "Resfresh App"
     )
  )
  server <- function(input, output, session) {
    observeEvent(input$refreshApp, {
      browsertools::refresh_page()
    })
  }
  shinyApp(ui, server)
}
}
\keyword{browsertools}
\keyword{page}
\keyword{refresh}
