% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_css.R
\name{remove_css}
\alias{remove_css}
\title{\code{remove_css}}
\usage{
remove_css(elem, css)
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{css}{a string containing the class(es) to remove from
an html element}
}
\value{
Removes a css class(es) to an element using id or classname.
}
\description{
Removes a css class(es) to an element using id or classname.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$head(
       tags$style(
          ".blue-text {
            color: blue;
          }"
       )
    ),
    tags$main(
       tags$h2("Remove CSS Example"),
       tags$p(
         id = "my-text-example",
         class = "blue-text",
         "Click the button below to remove the blue color from the text"
       ),
       tags$button(
          id = "removeCSS",
          class = "shiny-bound-input action-button",
          "Remove CSS"
       )
    )
  )
  server <- function(input, output, session) {
     observeEvent(input$removeCSS, {
       browsertools::remove_css(
         elem = "#my-text-example",
         css = "blue-text"
       )
     })
  }
  shinyApp(ui, server)
}
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/DOMTokenList/remove}
}
\seealso{
\code{\link[=add_css]{add_css()}}, \code{\link[=toggle_css]{toggle_css()}}
}
\keyword{browsertools}
\keyword{css}
\keyword{remove}
