% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{reset_input_groups}
\alias{reset_input_groups}
\title{\code{reset_input_groups}}
\usage{
reset_input_groups(elem, type = "all", default_val = FALSE)
}
\arguments{
\item{elem}{an ID of an input group to be reset (optional)}

\item{type}{the name of input type ("all", "radio", "checkbox") (optional)}

\item{default_val}{a logical value to use as the default or a name
of an html attribute to use a the default value.
This is used in \code{elem.checked = ...}
(optional)}
}
\value{
Reset input group values to default
}
\description{
Resets the values of all input groups by ID or type
}
\keyword{browsertools,}
\keyword{groups}
\keyword{input}
\keyword{inputs,}
\keyword{reset,}
