% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{run_demo}
\alias{run_demo}
\title{\code{run_demo}}
\usage{
run_demo()
}
\value{
Start the browsertools demo app
}
\description{
Start the browsertools demo app
}
\examples{
browsertools::run_demo()
}
\keyword{browsertools}
\keyword{demo}
