% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{scroll_to_top}
\alias{scroll_to_top}
\title{\code{scroll_to_top}}
\usage{
scroll_to_top()
}
\value{
Scrolls the window to the top of the page
}
\description{
Scrolls the window to the top of the page
}
\examples{
scroll_to_top()
}
\keyword{browsertools,}
\keyword{scroll}
