% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{set_document_title}
\alias{set_document_title}
\title{\code{set_document_title}}
\usage{
set_document_title(title, append = FALSE)
}
\arguments{
\item{title}{a string containing a title of the document}

\item{append}{if TRUE the title will be added to the current title}
}
\description{
Set or change the document title
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Setting the Document Title"),
    tags$p("Enter a new title for the document"),
    tags$input(
      id = "titleInput",
      class = "shiny-bound-input",
      type = "text"
    ),
    tags$button(
       id = "submit",
       class = "shiny-bound-input action-button",
       "Submit"
    )
  )
  server <- function(input, output, session) {
    observeEvent(input$submit, {
       browsertools::set_document_title(
         title = as.character(input$titleInput)
       )
    })
  }
  shinyApp(ui, server)
}
}
\keyword{browsertools}
\keyword{document}
\keyword{title}
