% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_elem.R
\name{show_elem}
\alias{show_elem}
\title{\code{show_elem}}
\usage{
show_elem(elem)
}
\arguments{
\item{elem}{the id or class name of an html elem}
}
\value{
Show a hidden html element
}
\description{
Shows an html element by id or class name.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
      browsertools::use_browsertools(),
      tags$main(
          tags$h2("Show Element Example"),
          tags$p(
              id = "my-text-example",
              "Click the button below to show a hidden element."
          ),
          tags$button(
              id = "showElement",
              class = "shiny-bound-input action-button",
              "Show Element"
          ),
          browsertools::hidden(
              tags$p(
                  id = "hiddenMessage",
                  "You found the hidden message!"
              )
          )
      )
  )
  server <- function(input, output, session) {
      observeEvent(input$showElement, {
          browsertools::show_elem(
              elem = "#hiddenMessage"
          )
      })
  }
  shinyApp(ui, server)
}

}
\seealso{
\code{\link[=hide_elem]{hide_elem()}}, \code{\link[=hidden]{hidden()}}, \code{\link[=toggle_elem]{toggle_elem()}}
}
\keyword{browsertools}
\keyword{show}
