% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{toggle_elem}
\alias{toggle_elem}
\title{\code{toggle_elem}}
\usage{
toggle_elem(elem, css = "browsertools-hidden")
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{css}{a string containing the class to remove from an html element
(default: browsertools-hidden; package default)}
}
\value{
Toggle the visibility of an html element
}
\description{
Toggle the visibility of an html element
}
\examples{
toggle_elem(elem = "#mydiv")
}
\keyword{browsertools}
\keyword{element}
\keyword{toggle}
