% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{use_browsertools}
\alias{use_browsertools}
\title{\code{use_browsertools}
Function that loads all assets into your shiny app}
\usage{
use_browsertools()
}
\value{
Function that loads all assets into your shiny app
}
\description{
\code{use_browsertools}
Function that loads all assets into your shiny app
}
\examples{
browsertools::use_browsertools()
}
\keyword{browsertools,}
\keyword{use}
