% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_css.R
\name{add_css}
\alias{add_css}
\title{Add CSS}
\usage{
add_css(elem, css)
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{css}{a string or character array containing css classes}
}
\value{
Adds a css class(es) to an element using id or classname.
}
\description{
Adds a css class(es) to an element using id or classname.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$head(
       tags$style(
          ".blue-text {
            color: blue;
          }",
         ".font-size-lg" {
            font-size: 150\%;
         }
       )
    ),
    tags$main(
       tags$h2("Add CSS Example"),
       tags$p(
         id = "my-text-example",
         "Click the button below to change the text color to blue"
       ),
       tags$button(
          id = "addCSS",
          class = "shiny-bound-input action-button",
          "Add CSS"
       )
    )
  )
  server <- function(input, output, session) {
     observeEvent(input$addCSS, {
       browsertools::add_css(
         elem = "#my-text-example",
         css = c("blue-text", "font-size-lg")
       )
     })
  }
  shinyApp(ui, server)
}
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/DOMTokenList/add}
}
\seealso{
\code{\link[=remove_css]{remove_css()}}, \code{\link[=toggle_css]{toggle_css()}}
}
\keyword{add}
\keyword{browsertools}
\keyword{css}
