% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{add_css}
\alias{add_css}
\title{\code{add_css}}
\usage{
add_css(elem, css, session = getDefaultReactiveDomain())
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{css}{a string containing the class(es) to add to
an html element}

\item{session}{a shiny environment}
}
\value{
Adds a css class(es) to an element using id or classname.
}
\description{
Adds a css class(es) to an element using id or classname.
}
\examples{
add_css(elem = "#mydiv", css = "some-css-class")
}
\keyword{add}
\keyword{browsertools,}
\keyword{css,}
