% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{alert}
\alias{alert}
\title{\code{alert}}
\usage{
alert(message)
}
\arguments{
\item{message}{a message to send}
}
\value{
Send an alert dialog to the browser
}
\description{
Send an alert dialog to the browser
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Display an alert"),
    tags$p("Click the button below to display an alert."),
    tags$button(
       id = "alert",
       class = "shiny-bound-input action-button",
       "Display Alert"
    )
  )
  server <- function(input, output, session) {
    observeEvent(input$alert, {
        browsertools::alert(
            message = "This is an alert"
        )
    })
  }
  shinyApp(ui, server)
}
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Window/alert}
}
\keyword{alert}
\keyword{browsertools,}
