% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_js_object.R
\name{as_js_object}
\alias{as_js_object}
\title{\code{as_js_object}}
\usage{
as_js_object(x)
}
\arguments{
\item{x}{a data frame object}
}
\value{
Restructures a data.frame to JavaScript Object
}
\description{
Restructure a data.frame to JavaScript Object
}
\examples{
if (interactive()) {
 library(shiny)
 ui <- tagList(
     browsertools::use_browsertools(),
     tags$p("Open the browser's dev console to view the data")
 )
 server <- function(input, output, session) {
     n <- 10
     df <- data.frame(
         group = sample(letters, n),
         x = rnorm(n),
         y = rnorm(n),
         z = rnorm(n)
     )
     browsertools::console_table(data = df)
  }
  shinyApp(ui, server)
}

}
\seealso{
\code{\link[=console_log]{console_log()}}, \code{\link[=console_table]{console_table()}}
}
\keyword{browsertools}
\keyword{object}
