% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{console_error}
\alias{console_error}
\title{\code{console_error}}
\usage{
console_error(message)
}
\arguments{
\item{message}{an error to display}
}
\value{
Sends an error message to the browser's console
}
\description{
Sends an error message to the brower's console
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Display an Error Message in the Dev Console"),
    tags$p("Open the browser's dev console and click the button below."),
    tags$button(
       id = "sendError",
       class = "shiny-bound-input action-button",
       "Send Error"
    )
  )
  server <- function(input, output, session) {
    observeEvent(input$sendError, {
        browsertools::console_error(
            message = "This is an error message"
        )
    })
  }
  shinyApp(ui, server)
}
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Console/error}
}
\keyword{browsertools}
\keyword{console}
\keyword{debugging}
