% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{console_log}
\alias{console_log}
\title{\code{console_log}}
\usage{
console_log(message, expand = FALSE)
}
\arguments{
\item{message}{a message to display}

\item{expand}{an option to expand arrays and objects (default: false)}
}
\value{
Outputs an object to the browser's console
}
\description{
Outputs an object to the browser's console
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Display Message in the Dev Console"),
    tags$p("Open the browser's dev console and click the button below."),
    tags$button(
       id = "sendMessage",
       class = "shiny-bound-input action-button",
       "Send Message"
    )
  )
  server <- function(input, output, session) {
    observeEvent(input$sendMessage, {
        browsertools::console_log(
            message = "This is a message"
        )
    })
  }
  shinyApp(ui, server)
}
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Console/log}
}
\keyword{browsertools}
\keyword{console}
\keyword{debugging}
