% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{console_log}
\alias{console_log}
\title{\code{console_log}}
\usage{
console_log(message)
}
\arguments{
\item{message}{a message to display}
}
\value{
Outputs an object to the browser's console
}
\description{
Outputs an object to the browser's console
}
\examples{
console_log(x = "Hello, world!")
}
\keyword{browsertools,}
\keyword{console}
\keyword{debugging,}
