% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{console_table}
\alias{console_table}
\title{\code{console_table}}
\usage{
console_table(data, to_json = TRUE)
}
\arguments{
\item{data}{an object to display in the browser (data.frame, etc.)}

\item{to_json}{an optional argument that transforms input data to json}
}
\value{
Outputs an object to the browser's console
}
\description{
Outputs an data object to the browser's console in table format
}
\examples{
if (interactive()) {
 library(shiny)
 ui <- tagList(
     browsertools::use_browsertools(),
     tags$p("Open the browser's dev console")
 )
 server <- function(input, output, session) {
     n <- 10
     df <- data.frame(
         group = sample(letters, n),
         x = rnorm(n),
         y = rnorm(n),
         z = rnorm(n)
     )
     browsertools::console_table(data = df)
  }
  shinyApp(ui, server)
}
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Console/table}
}
\keyword{browsertools}
\keyword{console}
\keyword{debugging}
