% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{console_table}
\alias{console_table}
\title{\code{console_table}}
\usage{
console_table(data, to_json = TRUE)
}
\arguments{
\item{data}{an object to display in the browser (data.frame, etc.)}

\item{to_json}{an optional argument that transforms input data to json}
}
\value{
Outputs an object to the browser's console
}
\description{
Outputs an data object to the browser's console in table format
}
\examples{
console_table(data = iris)
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Console/table}
}
\keyword{browsertools,}
\keyword{console}
\keyword{debugging,}
