% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_table.R
\name{console_table}
\alias{console_table}
\title{\code{console_table}}
\usage{
console_table(data, to_json = TRUE)
}
\arguments{
\item{data}{an object to display in the browser (data.frame, etc.)}

\item{to_json}{If TRUE, data will be converted to a JS object}
}
\value{
Outputs an object to the browser's console
}
\description{
Display data in the browser's console as a data table. By default, this
function transforms input data object as a JavaScript function using the
function \code{as_js_object} (also included in this package). Alternatively,
you can pass in a list object and set \code{to_json} to \code{FALSE}.
}
\examples{
if (interactive()) {
 library(shiny)
 ui <- tagList(
     browsertools::use_browsertools(),
     tags$p("Open the browser's dev console")
 )
 server <- function(input, output, session) {
     n <- 10
     df <- data.frame(
         group = sample(letters, n),
         x = rnorm(n),
         y = rnorm(n),
         z = rnorm(n)
     )
     browsertools::console_table(data = df)
  }
  shinyApp(ui, server)
}

}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Console/table}
}
\seealso{
\code{\link[=as_js_object]{as_js_object()}}, \code{\link[=console_log]{console_log()}}
}
\keyword{browsertools}
\keyword{console}
\keyword{debugging}
