% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{console_warn}
\alias{console_warn}
\title{\code{console_warn}}
\usage{
console_warn(message, session = getDefaultReactiveDomain())
}
\arguments{
\item{message}{a message to display}

\item{session}{a shiny environment}
}
\value{
Outputs a warning message to the console
}
\description{
Outputs a warning message to the console
}
\examples{
console_warn(message = "this is a warning message")
}
\keyword{browsertools,}
\keyword{console,}
\keyword{warn}
