% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug}
\alias{debug}
\title{\code{debug}}
\usage{
debug()
}
\value{
Print JavaScript errors in the R console
}
\description{
Print JavaScript errors in the R console. This may be useful for debugging
errors that may arise when using functions included in this pacakge. For
example, if an element cannot be found using the \code{add_css} function, the
\code{debug} function will print the corresponding error. (The correct selector
for \code{add_css} is \verb{#txt}.)
}
\examples{
if (interactive()) {
  ui <- tagList(
    tags$head(
      tags$style(
        ".blue-text {
          color: blue;
        }"
      )
    ),
    tags$h2("Browsertools demo"),
    tags$p(
      id = "txt",
      "JavaScript messages will print in the R Console"
    )
  )
  server <- function(input, output, session) {
    browsertools::debug()
    browsertools::add_css(
      elem = "#t",
      css = "blue-text"
    )
  }
  shinyApp(ui, server)
}

}
\seealso{
\code{\link[=print_elem]{print_elem()}}
}
\keyword{browsertools}
\keyword{debug}
\keyword{error}
