% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug}
\alias{debug}
\title{\code{Debug}}
\usage{
debug()
}
\value{
Print JavaScript errors in the R console
}
\description{
Print JavaScript errors in the R console. This function
is designed to be used within the Shiny Server
}
\examples{
browsertools::debug()
}
\keyword{browsertools,}
\keyword{debug,}
\keyword{error}
