% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{hidden}
\alias{hidden}
\title{\code{hidden}}
\usage{
hidden(..., css = "browsertools-hidden")
}
\arguments{
\item{...}{A series HTML elements}

\item{css}{An optional class name used to hide elements
(Default class = "browsertools")}
}
\value{
Hide elements by when rendered
}
\description{
Hide elements by when rendered
}
\examples{
if (interactive()) {
 hidden(tags$p("hello, world"), tags$p("this is a test"))
}
}
\keyword{browsertools,}
\keyword{hidden}
