% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_elem.R
\name{hide_elem}
\alias{hide_elem}
\title{\code{hide_elem}}
\usage{
hide_elem(elem)
}
\arguments{
\item{elem}{the id or class name of an html elem to hide}
}
\value{
hide an HTML element
}
\description{
Hides an html element by id or class name.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
      browsertools::use_browsertools(),
      tags$main(
          tags$h2("Hide Element Example"),
          tags$p(
              id = "my-text-example",
              "Click the button below to hide the message below."
          ),
          tags$button(
              id = "hideElement",
              class = "shiny-bound-input action-button",
              "Hide Element"
          ),
          tags$p(
              id = "myMessage",
              "Hide this message!"
          )
      )
  )
  server <- function(input, output, session) {
      observeEvent(input$hideElement, {
          browsertools::hide_elem(
              elem = "#myMessage"
          )
      })
  }
  shinyApp(ui, server)
}

}
\seealso{
\code{\link[=show_elem]{show_elem()}}, \code{\link[=toggle_elem]{toggle_elem()}}, \code{\link[=hidden]{hidden()}}
}
\keyword{browsertools}
\keyword{hide}
