% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{hide_elem}
\alias{hide_elem}
\title{\code{hide_elem}}
\usage{
hide_elem(elem, css = "browsertools-hidden")
}
\arguments{
\item{elem}{the id or class name of an html elem}

\item{css}{a string containing the class to remove from an html element
(default class is \code{hidden})}
}
\description{
Hides an html element by id or class name. This function, by defauly,
adds the class \code{hidden} from the desired element and updates
aria attriute \code{hidden}. In your css file, create a corresponding
selector called \code{hidden} and define the hide styles. This allows you
to add delays and apply other transformations.
}
\examples{
hide_elem(elem = "#mydiv")
hide_elem(elem = "#mydiv", css = "show-div")
}
\keyword{browsertools,}
\keyword{css,}
\keyword{show}
