% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{inner_html}
\alias{inner_html}
\title{\code{inner_html}}
\usage{
inner_html(elem, string, delay = NULL)
}
\arguments{
\item{elem}{the ID or class name of an html element}

\item{string}{a string to display in the element}

\item{delay}{an optional arg to add a brief pause
before sending the content to the html element.
Ideal for content that is rendered by the server.
Input is time in milliseconds.}
}
\value{
Update the content of an element by id or class name
}
\description{
Update the content of an element by id or class name
}
\keyword{browsertools,}
\keyword{html}
\keyword{inner,}
