% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_text.R
\name{inner_text}
\alias{inner_text}
\title{\code{inner_text}}
\usage{
inner_text(elem, content, append = FALSE, delay = NULL)
}
\arguments{
\item{elem}{an element to select (e.g., ID, class, tag, etc.)}

\item{content}{content to insert}

\item{append}{an option that appends value of string or replaces it}

\item{delay}{an optional arg that adds a brief pause before
sending the content to the html element. Ideal
for content that is rendered server-side. Input
time is in milliseconds.}
}
\value{
Modify the text of an element
}
\description{
Modify the text of an element
}
\examples{
if (interactive()) {
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Inner Html Example"),
    tags$p("Click the button to update the following message."),
    tags$p(
      id = "textToUpdate",
      "[This text will be updated]"
    ),
    tags$button(
      id = "updateText",
      class = "shiny-bound-input action-button",
      "Update Text"
    )
  )
  server <- function(input, output, session) {
    browsertools::inner_text(
      elem = "#textToUpdate",
      content = "This is a new message!"
    )
  }
  shinyApp(ui, server)
}

}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/innerText}
}
\seealso{
\code{\link[=inner_html]{inner_html()}}, \code{\link[=insert_adjacent_html]{insert_adjacent_html()}}
}
\keyword{browsertools}
\keyword{innertext}
