% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_adjacent_html.R
\name{insert_adjacent_html}
\alias{insert_adjacent_html}
\title{\code{insert_adjacent_html}}
\usage{
insert_adjacent_html(id, content, position = "beforeend")
}
\arguments{
\item{id}{an id of the parent element to render new elements into}

\item{content}{An html string, shiny tag, or shiny tag list}

\item{position}{a position relative to the reference element
(i.e., \code{beforebegin}, \code{afterbegin}, \code{beforeend},
or \code{afterend})}
}
\value{
Create a new html element using shiny tags as a child of an element
}
\description{
Create and render a new html element(s) using shiny tags. Content is added
to the document using a reference container (i.e., body, div, etc.). Use
the argument \code{position} to specify where the content should be added.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$main(
      id = "main",
      tags$h1("Inner Adjacent HTML Example")
    )
  )
  server <- function(input, output, session) {
  insert_adjacent_html(
    id = "main",
    content = tagList(
        tags$h2("Hello, world!"),
        tags$p("How are you doing today?")
    ),
    position = "afterbegin"
  )
}
}

}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Element/insertAdjacentHTML}
}
\seealso{
\code{\link[=inner_text]{inner_text()}}, \code{\link[=inner_html]{inner_html()}}
}
\keyword{browsertools}
\keyword{html}
\keyword{insertHTML}
