% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{insert_adjacent_html}
\alias{insert_adjacent_html}
\title{\code{insert_adjacent_html}}
\usage{
insert_adjacent_html(
  id,
  html,
  position = "beforeend",
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{an id of the parent element to render new elements into}

\item{html}{an html tag or list object containing elements to render}

\item{position}{a position relative to the html element
(i.e., \code{beforebegin}, \code{afterbegin}, \code{beforeend},
or \code{afterend})}

\item{session}{a shiny environment}
}
\value{
Create a new html element using shiny tags as a child of an element
}
\description{
Create a new html element using shiny tags as a child of an element
}
\keyword{browsertoos,}
\keyword{html}
