% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{insert_adjacent_html}
\alias{insert_adjacent_html}
\title{\code{insert_adjacent_html}}
\usage{
insert_adjacent_html(id, html, position = "beforeend")
}
\arguments{
\item{id}{an id of the parent element to render new elements into}

\item{html}{an html tag or list object containing elements to render}

\item{position}{a position relative to the html element
(i.e., \code{beforebegin}, \code{afterbegin}, \code{beforeend},
or \code{afterend})}
}
\value{
Create a new html element using shiny tags as a child of an element
}
\description{
Create a new html element using shiny tags as a child of an element
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$main(
      id = "main",
      tags$h1("Inner Adjacent HTML Example")
    )
  )
  server <- function(input, output, session) {
  insert_adjacent_html(
    id = "main",
    html = tagList(
        tags$h2("Hello, world!"),
        tags$p("How are you doing today?")
    ),
    position = "afterbegin"
  )
}
}
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Element/insertAdjacentHTML}
}
\keyword{browsertools}
\keyword{html}
\keyword{insertHTML}
