% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_elem.R
\name{print_elem}
\alias{print_elem}
\title{\code{print_elem}}
\usage{
print_elem(elem)
}
\arguments{
\item{elem}{an element to select (i.e., ID, class, tag, etc.)}
}
\value{
Select and print an element in both the JavaScript and R console
}
\description{
Select and print an html element in both the JavaScript and R consoles
}
\examples{
print_elem(elem = "#mydiv")
}
\keyword{browsertools,}
\keyword{print}
