% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_elem.R
\name{print_elem}
\alias{print_elem}
\title{\code{print_elem}}
\usage{
print_elem(elem)
}
\arguments{
\item{elem}{an element to select (i.e., ID, class, tag, etc.)}
}
\value{
Select and print an element in both the JavaScript and R console
}
\description{
Select and print an html element in both the JavaScript and R consoles
}
\examples{
if (interactive()) {
   ui <- tagList(
       browsertools::use_browsertools(),
       tags$div(
         id = "mydiv",
         tags$h2("My Element"),
         tags$p(
           "This is an example element. The structure will be printed",
           "in the R console."
         )
       )
   )
   server <- function(input, output) {
     observe({
        print_elem(elem = "#mydiv")
     })
   }
}
}
\keyword{browsertools,}
\keyword{print}
