% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{refresh_page}
\alias{refresh_page}
\title{\code{refresh_page}}
\usage{
refresh_page()
}
\value{
Trigger a page refres
}
\description{
Trigger a page refresh
}
\keyword{browsertools,}
\keyword{page,}
\keyword{refresh}
