% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{remove_css}
\alias{remove_css}
\title{\code{remove_css}}
\usage{
remove_css(elem, css)
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{css}{a string containing the class(es) to remove from
an html element}
}
\value{
Removes a css class(es) to an element using id or classname.
}
\description{
Removes a css class(es) to an element using id or classname.
}
\examples{
remove_css(elem = "#mydiv", css = "some-css-class")
}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/DOMTokenList/remove}
}
\keyword{browsertools,}
\keyword{css,}
\keyword{remove}
