% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{remove_element}
\alias{remove_element}
\title{\code{remove_element}}
\usage{
remove_element(elem, session = getDefaultReactiveDomain())
}
\arguments{
\item{elem}{an ID of the element to remove}

\item{session}{a shiny environment}
}
\value{
Removes an html element from the DOM
}
\description{
Removes an html element from the DOM
}
\examples{
remove_element(elem = "#mydiv")
}
\keyword{browsertools,}
\keyword{element}
\keyword{remove,}
