% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_element.R
\name{remove_element}
\alias{remove_element}
\title{\code{remove_element}}
\usage{
remove_element(elem)
}
\arguments{
\item{elem}{an ID of the element to remove}
}
\value{
Removes an html element from the DOM
}
\description{
Removes an html element from the DOM
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Remove Element Example"),
    tags$p(
      id = "textToRemove",
      "Click the button below to remove this message."
    ),
    tags$button(
      id = "removeMessage",
      class = "shiny-bound-input action-button",
      "Remove Message"
    )
  )
  server <- function(input, output, session) {
    observeEvent(input$removeMessage, {
      browsertools::remove_element(
        elem = "#textToRemove"
      )
    })
  }
  shinyApp(ui, server)
}

}
\seealso{
\code{\link[=inner_html]{inner_html()}}, \code{\link[=insert_adjacent_html]{insert_adjacent_html()}}
}
\keyword{browsertools}
\keyword{element}
\keyword{remove}
