% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{remove_element_attribute}
\alias{remove_element_attribute}
\title{\code{remove_element_attribute}}
\usage{
remove_element_attribute(elem, attr, session = getDefaultReactiveDomain())
}
\arguments{
\item{elem}{an ID of the html element to be modified}

\item{attr}{the name of the attribute to remove}

\item{session}{a shiny environment}
}
\value{
Remove an attribute from an html element
}
\description{
Removes an html attribute from an element
}
\examples{
remove_element_attribute(elem = "#mydiv", attr = "class")
}
\keyword{attribute}
\keyword{browsertools,}
\keyword{remove,}
