% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{scroll_to}
\alias{scroll_to}
\title{\code{scroll_to}}
\usage{
scroll_to(x = 0, y = 0)
}
\arguments{
\item{x}{amount (in pixels) to scroll along the horizontal axis
starting from the top left (default: 0)}

\item{y}{amount (in pixels) to scroll along the vertical axis
starting from the top left (default: 0)}
}
\value{
Scrolls the window to the top of the page or a user
         defined coordinates
}
\description{
Scrolls the window to the top of the page or a user defined coordinates
}
\examples{
scroll_to()
scroll_to(y = 250)
}
\keyword{browsertools,}
\keyword{scroll}
