% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{scroll_to_top}
\alias{scroll_to_top}
\title{\code{scroll_to_top}}
\usage{
scroll_to_top()
}
\value{
Adds a css class(es) to an element using id or classname.
}
\description{
Moves the current viewport to the top of the page
}
\examples{
scroll_to_top()
}
\keyword{browsertools,}
\keyword{scroll}
