% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{set_element_attribute}
\alias{set_element_attribute}
\title{\code{set_element_attribute}}
\usage{
set_element_attribute(elem, attr, value)
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{attr}{the name of the attribute to update}

\item{value}{the value to add to the attribute}
}
\value{
Set an attribute of an html element
}
\description{
Set an attribute of an html element
}
\examples{
set_element_attribute(elem = "#mydiv", attr = "data-value", value = "12345")
}
\keyword{attribute,}
\keyword{browsertools,}
\keyword{value}
