% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{set_element_attribute}
\alias{set_element_attribute}
\title{\code{set_element_attribute}}
\usage{
set_element_attribute(elem, attr, value)
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{attr}{the name of the attribute to update}

\item{value}{the value to add to the attribute}
}
\value{
Set an attribute of an html element
}
\description{
Set an attribute of an html element
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$head(
      tags$style(
        "[aria-hidden='true'] {
             position: absolute;
             width: 1px;
             height: 1px;
             margin: -1px;
             clip: rect(0, 0, 0, 0);
             clip: rect(0 0 0 0);
             overflow: hidden;
             white-space: nowrap;
        }"
      )
    ),
    tags$h2("Set Element Attribute Example"),
    tags$p(
      "Click the button below to set an attribute of an element"
    ),
    tags$button(
      id = "setAttr",
      class = "shiny-bound-input action-button",
      "Set Attribute"
    ),
    tags$p(
      id = "mytext",
      `aria-hidden` = "true",
      "This element will be modified."
    )
  )
  server <- function(input, output, session) {
    observeEvent(input$removeAttr, {
      browsertools::set_element_attribute(
        elem = "#mytext",
        attr = "aria-hidden",
        value = "true"
      )
    })
  }
  shinyApp(ui, server)
}
}
\keyword{attribute}
\keyword{browsertools}
\keyword{value}
