% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{show_elem}
\alias{show_elem}
\title{\code{show_elem}}
\usage{
show_elem(elem, css = "browsertools-hidden")
}
\arguments{
\item{elem}{the id or class name of an html elem}

\item{css}{a string containing the class to remove from an html element
(default class is \code{hidden})}
}
\description{
Shows an html element by id or class name. This function, by defauly,
removes the class \code{hidden} from the desired element and updates
aria attriute \code{hidden}. In your css file, create a corresponding
selector called \code{hidden} and define the show styles. This allows you
to add delays and apply other transformations.
}
\examples{
show_elem(elem = "#mydiv")
show_elem(elem = "#mydiv", css = "show-div")
}
\keyword{browsertools,}
\keyword{css,}
\keyword{show}
