% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{show_elem}
\alias{show_elem}
\title{\code{show_elem}}
\usage{
show_elem(elem, css = "browsertools-hidden")
}
\arguments{
\item{elem}{the id or class name of an html elem}

\item{css}{a string containing the class to remove from an html element
(default class is \code{hidden})}
}
\description{
Shows an html element by id or class name. This function, by defauly,
removes the class \code{hidden} from the desired element and updates
aria attriute \code{hidden}. In your css file, create a corresponding
selector called \code{hidden} and define the show styles. This allows you
to add delays and apply other transformations.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
      browsertools::use_browsertools(),
      tags$main(
          tags$h2("Show Element Example"),
          tags$p(
              id = "my-text-example",
              "Click the button below to show a hidden element."
          ),
          tags$button(
              id = "showElement",
              class = "shiny-bound-input action-button",
              "Show Element"
          ),
          browsertools::hidden(
              tags$p(
                  id = "hiddenMessage",
                  "You found the hidden message!"
              )
          )
      )
  )
  server <- function(input, output, session) {
      observeEvent(input$showElement, {
          browsertools::show_elem(
              elem = "#hiddenMessage"
          )
      })
  }
  shinyApp(ui, server)
}
}
\keyword{browsertools}
\keyword{css}
\keyword{show}
