% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{toggle_css}
\alias{toggle_css}
\title{\code{toggle_css}}
\usage{
toggle_css(elem, css)
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{css}{a string containing the class to remove from
an html element}
}
\value{
Toggles a css state of an html element
}
\description{
Toggles a css state of an html element
}
\examples{
toggle_css(elem = "#mydiv", css = "mytheme")
}
\keyword{attribute,}
\keyword{browsertools,}
\keyword{value}
