% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_css.R
\name{toggle_css}
\alias{toggle_css}
\title{\code{toggle_css}}
\usage{
toggle_css(elem, css)
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{css}{a string or character array containing css classes}
}
\value{
Toggles a css state of an html element
}
\description{
Toggles a css state of an html element
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$head(
       tags$style(
          ".blue-text {
            color: blue;
          }"
       )
    ),
    tags$main(
       tags$h2("Toggle CSS Example"),
       tags$p(
         id = "my-text-example",
         "Click the button below to toggle the text color."
       ),
       tags$button(
          id = "toggleCSS",
          class = "shiny-bound-input action-button",
          "Toggle CSS"
       )
    )
  )
  server <- function(input, output, session) {
     observeEvent(input$toggleCSS, {
       browsertools::toggle_css(
         elem = "#my-text-example",
         css = "blue-text"
       )
     })
  }
  shinyApp(ui, server)
}

}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/DOMTokenList/toggle}
}
\seealso{
\code{\link[=remove_css]{remove_css()}}, \code{\link[=add_css]{add_css()}}
}
\keyword{browsertools}
\keyword{css}
\keyword{toggle}
