% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_elem.R
\name{toggle_elem}
\alias{toggle_elem}
\title{\code{toggle_elem}}
\usage{
toggle_elem(elem, css = "browsertools-hidden")
}
\arguments{
\item{elem}{the id or class name of an html element}

\item{css}{a string containing the class to remove from an html element
(default: browsertools-hidden; package default)}
}
\value{
Toggle the visibility of an html element
}
\description{
Toggle the visibility of an html element
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
      browsertools::use_browsertools(),
      tags$main(
          tags$h2("Toggle Element Example"),
          tags$p(
              id = "my-text-example",
              "Click the button below to toggle the hidden element."
          ),
          tags$button(
              id = "toggleElement",
              class = "shiny-bound-input action-button",
              "Toggle Element"
          ),
          browsertools::hidden(
              tags$p(
                  id = "hiddenMessage",
                  "You found the hidden message!"
              )
          )
      )
  )
  server <- function(input, output, session) {
      observeEvent(input$toggleElement, {
          browsertools::toggle_elem(
              elem = "#hiddenMessage"
          )
      })
  }
  shinyApp(ui, server)
}

}
\seealso{
\code{\link[=hide_elem]{hide_elem()}}, \code{\link[=show_elem]{show_elem()}}
}
\keyword{browsertools}
\keyword{element}
\keyword{toggle}
