% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_browsertools.R
\name{use_browsertools}
\alias{use_browsertools}
\title{\code{use_browsertools}}
\usage{
use_browsertools()
}
\value{
Function that loads all assets into your shiny app
}
\description{
Function that loads the browsertools JavaScript file into your shiny app.
This function is required and should be called at the top of the
Shiny app. If you are using other ShinyUI layouts, you may need to wrap
your app in \code{tagList}. (The size of the js file is approximately 5.1kb.)
}
\examples{
if (interactive()) {
  ui <- tagList(
    browsertools::use_browsertools(),
    tags$h2("Hello, world!")
  )
  server <- function(input, output, session) {
  }
  shinyApp(ui, server)
}

}
\keyword{browsertools}
\keyword{initial}
\keyword{setup}
\keyword{use}
